--- .gitignore
@@ -1,3 +1,0 @@
-cmake-build-release
-cmake-build-debug
-.idea
